consolidate exporter code
